test: reorganize idle lifecycle fixtures
