+++ crates/runner/src/idle_pool.rs
@@ -993,6 +993,9 @@ pub enum ParkResult {
 }
 
 #[cfg(test)]
+mod destroy_tests;
+
+#[cfg(test)]
 mod tests {
     use super::*;
 
@@ -1048,33 +1051,6 @@ mod tests {
         }
     }
 
-    async fn make_idle_destroy_payload(overrides: Arc<MockSandboxOverrides>) -> IdleDestroyPayload {
-        let sandbox_id = SandboxId::new_v4();
-        let factory: Arc<Box<dyn SandboxFactory>> = Arc::new(Box::new(
-            MockSandboxFactory::with_overrides(Arc::clone(&overrides)),
-        ));
-        let sandbox = factory
-            .create(SandboxConfig {
-                id: sandbox_id,
-                resources: ResourceLimits {
-                    cpu_count: 2,
-                    memory_mb: 4096,
-                },
-                device_rate_limits: None,
-                workspace_drive: None,
-            })
-            .await
-            .expect("create sandbox");
-
-        IdleDestroyPayload {
-            resources: IdleSandboxResources {
-                sandbox,
-                factory,
-                workspace_promotion: None,
-            },
-        }
-    }
-
     async fn make_idle_park_request(
         overrides: Arc<MockSandboxOverrides>,
         session_id: &str,
@@ -1107,32 +1083,6 @@ mod tests {
             storage_fingerprints: StorageFingerprints::default(),
             workspace_promotion: None,
         })
-    }
-
-    #[tokio::test]
-    async fn idle_destroy_payload_stop_error_completes_after_destroy() {
-        let overrides = Arc::new(MockSandboxOverrides::new());
-        overrides.push_stop_result(Err(sandbox::SandboxError::Start {
-            message: "simulated idle stop failure".into(),
-        }));
-        let payload = make_idle_destroy_payload(Arc::clone(&overrides)).await;
-
-        let outcome = payload.stop_and_destroy().await;
-
-        assert_eq!(outcome, DestroyOutcome::Completed);
-        assert_eq!(overrides.destroy_call_count(), 1);
-    }
-
-    #[tokio::test]
-    async fn idle_destroy_payload_stop_panic_is_uncertain_after_destroy() {
-        let overrides = Arc::new(MockSandboxOverrides::new());
-        overrides.push_stop_panic("simulated idle stop panic");
-        let payload = make_idle_destroy_payload(Arc::clone(&overrides)).await;
-
-        let outcome = payload.stop_and_destroy().await;
-
-        assert_eq!(outcome, DestroyOutcome::Uncertain);
-        assert_eq!(overrides.destroy_call_count(), 1);
     }
 
     #[tokio::test]
